[maze] Major refacotr, adding polar grids
This is a massive refactor away from the use of *BaseCell pointers in
favor of the use of a Cell interface, which abstracts away the specific
type so that we can use *PolarCell pointers in the implementation of
PolarGrids.

It also includes much of the ground work necessary for polar grids,
although it's not all the way there yet.