[maze] Major refactor
This is a really comprehensive refactor.

Changes:

1. In order to support polar grids, I modified the inheritance model of
cells so that there's a baseCell, a GridCell (used for rectangular
grids) and a PolarCell (used for polar grids). GridCells and PolarCells
inherit from baseCells -- the only thing baseCells really do is to
provide a row and column field and associated getters for the Cell
interface.

2. Moved all of the cell-related logic out of the model package into the
model/cell package. I might move the other grid stuff into a model/grid
package, but I'm not in a rush to do that right now. It's a real mess to
refactor.

3. Finish adding support for Polar grids